requirements.txt
click
Flask
gunicorn
itsdangerous
Jinja2
MarkupSafe
Werkzeug
Flask-SocketIO
PyMySQL
spacy
spacy-legacy
spacy-loggers
numpy
scikit-learn
pandas
openai
regex
python-dotenv
gevent
Flask-Session
PyPDF2